// server/src/enums/FeatureAlias.ts
export enum FeatureAlias {
  SomeReviews = 'some.reviews',
  AllReviews = 'all.reviews',
  Filtering = 'filtering',
  Sorting = 'sorting',
  Search = 'search',
  Specializations = 'spec',
  WordCloud = 'word.cloud',
  Metrics = 'metrics',
  Trends = 'trends',
  EarlyAccess = 'roadmap',
  Reactions = 'reactions',
  Notifications = 'notifs',
  RepliesComments = 'replies.comments',
  Notes = 'notes',
  Enrollments = 'enrollments',
  Grades = 'grades',
  NLP = 'nlp',
  WhatIf = 'what_if',
  Crypto = 'crypto',
}